+++ app/view_models/v1/core_cost_summary.rb
@@ -1,5 +1,7 @@
 module V1
   class CoreCostSummary < BaseViewModel
+    include Shared::WorkItemSummary
+
     SKIPPED_TYPES = %w[travel waiting].freeze
 
     attribute :work_items
@@ -29,20 +31,8 @@ module V1
       @data_by_type ||= work_item_data + letter_and_call_data
     end
 
-    def work_item_data
-      work_items.map { |work_item| WorkItem.build_self(work_item) }
-                .group_by { |work_item| work_item.work_type.to_s }
-                .filter_map do |translated_work_type, work_items_for_type|
-                  work_type = work_items_for_type.first.work_type
-                  next if SKIPPED_TYPES.include?(work_type.value)
-
-                  # TODO: convert this to a time period to enable easy formating of output
-                  [
-                    translated_work_type,
-                    work_items_for_type.sum { |work_item| CostCalculator.cost(:work_item, work_item) },
-                    work_items_for_type.sum(&:time_spent),
-                  ]
-                end
+    def skip_work_item?(work_item)
+      SKIPPED_TYPES.include?(work_item.work_type.value)
     end
 
     def letter_and_call_data
